Migrate photoModal to it is respective view and draft explore view
